requirements.txt
fastapi
uvicorn[standard]
python-multipart
rembg[cpu]
slowapi
python-dotenv
pillow
numpy
opencv-python-headless
tqdm
requests
torch
torchvision